Began work on class Cache_Item and function InitCacheScan; finished function GenerateFileName
The class Cache_Item is CURRENTLY INCOMPLETE: the idea is to have parts of
the program that affects/manages ALL cached HTML as global functions,
while having the data and attributes of the cached HTML and
object-to-database communication and transactions be handled by the
attributes and methods of the class. Currently the __init__ constructor
is somewhat implemented. The function InitCacheScan will scan the database
file where each line in the file represents a cached html item.