Major filesystem, webpage, and Cache_Item management functionality added
Added global varaibles for the database text file: its name, folder,
content, and a dict with Cache_Item objects corresponding to each
entry/line. The Cache_Item class was renamed to Cache_Item_Object,
and its objects are now capable of reading and writing from their
corresponding HTML files, making webpage requests to update the
cache itself, and checking if the cached content is expired.
The code for checking an object's expiration used to be a
global function, but is now a class method, Implementation for
methods interacting with the database has started but is mostly
unfinished (borderline placeholders). Upon instantiation, a
Cache_Item object creates a placeholder HTML file. New global
functions GenerateCacheObject, WebpageFetch, DatabaseFileRead,
InitCacheScan, and InitFilesNeededCreate are partially to
mostly implemented: the code cannot restore Cache_Item objects
from the database as of now, but will in the future.