New garbage collection system, minor improvements to debug messages and DatabaseFileRead
The garbage collection of cached HTML files and objects in the database is
handled by the global function DatabaseCleanup. The function is intended to
be among the first functions called when fetcher.py is executed. The function
deletes expired objects instead of updating them like WebpageFetch normally
does. An object is considered "expired" if the current date-time exceeds
the expiration date-time. The procedure for trashing an expired database
entry is as follows: 1) remove the cached HTML file, 2) remove the entry
from the database (via cacheDatabaseModifiedLines), 3) perform a read-write
cycle of the database via DatabaseFileWriteReadCycle (MISTAKE: this should
be OUTSIDE of the FOR LOOP, THIS NEEDS TO BE FIXED LATER!), 4) updates
the line numbers for each cache object by calling DatabasePrepare for
each existing object in cacheReferenceDict.

In terms of the debug system, there are "ending" banners to function calls
alongside the starting banners so that debug information pertaining to a
specific function is not confused with other functions. Existing calls
to the debug printing functions throughout the program have been modified.
Honestly, the debug system is getting out of hand and is very difficult
to read: the DebugPrintFuncCall may be removed in favor of exclusively
singe-line printing.

DatabaseFileRead: slight parameter change from "initObjectCreate" to "init"
and checks to see if a line is populated by looking for tab characters.