Garbage collection bug fix: moved function calls outside of for loop
Moved the DatabaseFileWriteReadCycle and DatabasePrepare method calls to
appear OUTSIDE/AFTER the loop. This is because it is redundant to constantly
commit changes in the cacheDatabaseModifiedLines buffer every time that
buffer changes: commit when all the changes have been made and finalized.
Additional comments explaining the process have also been added.