Removed file and folder name lists in script: individual strings are now used.
Removed the fileNameList and folderNameList global list variables.
Updated global function InitFilesNeededCreate to now use the old individual
string variables and not the listts. My rationale: in the beginning I wanted
the program to be flexible and adapt to more folders and key files as needed,
but now I am confident that a single database file and a single folder for the
cached HTML is all I need.

IMPORTANT NOTE: additional testing for having different absolute and relative
paths versus the default is needed.