Major changes to WebpageFetch and ExpirationCheck; bug fix in DatabaseCleanup
Global function WebpageFetch now can accept non-default expiration times.
When the user specifies an expiration interval to WebpageFetch, it will only
be applied for brand new or expired (therefore updated) cache objects. New
behavior to parameters: 1) When parameter forceUpdateCache is True:
forcefully updates expiration interval with cacheObject.ExpirationIntervalChange.
2) When forceUpdateCache is False (normally retuning cached html):
the autoUpdateExpirationInterval is used with cacheObject.ExpirationCheck.
3) When new cache objects are created: the expirationIntervalMin parameter
is used with GenerateCacheObject.

Global function ExpirationCheck now can optionally, automatically update the
expiration interval when the cache object expires.

CRITICAL BUG with DatabaseCleanup: due to how the index for deleting lines
was determined with the cacheDatabaseReadLines variable INSTAD of the
cacheDatabaseModifiedLines, line deletions were inaccurate and did not
reflect staged changes to the database.