Changed comments, requirements, and removed scraper.py
Context: fetcher.py and the whole git repo was intended to be used for
an upcoming project to scrape and analyze the eBay website. However,
fetcher.py is now its own project and can be used across any web scraping
or communications project beyond eBay.

Therefore, the requirements.txt has been modified to remove any old
requirements for the eBay Scraper Project AND scraper.py was removed.

The requests library is NOT a part of the Standard Library: comments were
modified to reflect that.