Fetcher: by default, debug messages are not shown UNLESS Fetcher is ran directly (__main__).
It is NOT ideal for Fetcher to print debug messages in the console when Fetcher
is imported by other code: if needed, the user may enable the verbose printing.

The only downside is that for any code that imports Fetcher, they are only able
to enable the verbose printing AFTER fetcher is finished initializing (meaning
any debug messages or "soft" errors that occur during the initialization will NOT
appear in the console).

Perhaps in the future, a config file (or even in the DB file) could contain
settings and other flags for Fetcher to have persistent behavior set by devs
using Fetcher.