Critical bug fixes in Fetcher; created "tester" and "demo" scripts.
Two critical bug fixes in Fetcher:

1) Originally, in the constructor of a Cache Object, the HTML content
of the object is NOT updated whatsoever, even if the HTML content
is already cached in the database. This is a MISTAKE: the object
SHOULD immediately match its HTML content with its corresponding
cached HTML file immediately upon the object's
creation.

2) In the WebpageFetch function, if parameter forceUpdateCache were set
to True (meaning the user wants to forcefully update the cache, regardless
if it has expired or not), the program would crash due to how I forgot
to define cacheObject in the if-statement code block for forceUpdateCache.

The "tester" and "demo" scripts:

The "tester" is meant for me to test crucial features of fetcher before
finalizing its publication on GitHub (and that was how I found the two
bugs above).

The "demo" is meant to serve as an extremely simple script that people
may execute to see the core behavior of Fetcher from the perspective of
a developer wanting to import Fetcher for a project.