Demo script: all functionality now implemented; comment and printing cleanup later.
The demo script has been finished and tests the following features of Fetcher:

1) Before its expiration, a cached HTML will be used for a URL, meaning multiple
calls to WebpageFetch for that URL before expiration will yield the exact same
HTML (no new HTML is returned because the system has yet to actually communicate
with the website).

2) The expiration time for a particular URL can be changed with a call to
WebpageFetch with parameter "expirationIntervalMin" set.

3) Expiration time persistence; a cached HTML will expire with the expiration time
last set in WebpageFetch, meaning either the default expiration time or the
last change to the expiration time the user has made will persist without any
values passed into the parameter "expirationIntervalMin" until expirationIntervalMin
is not None for a future call to WebpageFetch.

4) A cached HTML can be updated before its expiration time by forcefully updating it
in a call to WebpageFetch with forceUpdateCache set to true.

Side note: the formatting for the printing is a little confusing and overwhelming.
I may simplify the printing for it to be more appaeling and immediately
undestandable. This may require changing the output of Fetcher.