demo.py: code cleanup complete, rearranged into phases with only one custom function.
All of the arbitrarily written code was standardized into one simple
function that is called 6 times (due to 6 phases). Each call to the
function is highly organized, readible, and supplemented with commentary in
the form of a "message" parameter, eliminating the need and dependence on
comments.

The demo script should be entirely finished at this point: no other
major improvements should be made.

Also removed the "ebay scraper" label in the main script.