Major changes: finished install+use in README; tester.py now has interface; fetcher.py behavior slightly altered
README: finished the "Installation and Basic Use" section of the README;
no images have been inserted yet, but will be placed soon at the top
of the README and under the "demo" and "tester" script section.
The "Rationale and Motvation" and "Fetcher in Action" are (hopefully) the
last major sections that need to be started.

TESTER: tester is no longer some trashy wrapper-like script for testing
fetcher behavior and functionality: it now features a proper text-based
user interface where a non-developer user can type in basic commands to tell
the tester script to call WebpageFetch on particular URLs.

FETCHER: now, when WebpageFetch is given an expiration interval, it WILL
update the expiration interval no matter what, regardless of whether or
not the cached HTML for the given URL has expired or not. Previously, the
HTML would need to be expired in that call in order for the expiration interval
to be updated.

DEMO: added note that webpages may not update during every single call to
WebpageFetch, therefore the HTML may be the "SAME" when it is expected to be
"DIFFERENT."