README: now finished; .github_readme assets folder added
The "Rationale and Motivation" and "Fetcher in Action" sections have been
completed. Screenshots have been added at the top and for the demo and
tester script sections. These screenshots are stored in the .github_readme
folder in the repo. This folder contains many unused screenshots that may
be deleted later to save on space; for now, these screenshots will remain
in case they are used later.